adjust base query structure

+++ domain-support/src/main/java/com/stellariver/milky/domain/support/event/FinalEventRouter.java
@@ -14,6 +14,6 @@ public @interface FinalEventRouter {
 
     String executor() default "asyncExecutor";
 
-    int order() default Integer.MAX_VALUE;
+    double order() default Double.MAX_VALUE;
 
 }
